Fixed Packet.GetBin reading length as short

+++ MabiPale2.Shared/Packet.cs
@@ -520,7 +520,7 @@ namespace MabiPale2.Shared
 				throw new ArgumentException("Expected Bin, got " + this.Peek() + ".");
 
 			_ptr += 1;
-			var len = IPAddress.NetworkToHostOrder(BitConverter.ToInt16(_buffer, _ptr));
+			var len = (int)(ushort)IPAddress.NetworkToHostOrder(BitConverter.ToInt16(_buffer, _ptr));
 			_ptr += 2;
 
 			var val = new byte[len];
